- Added more tools for managing experiments. - Added snippet and template "bank" for managing peersim configurations in a more reasonable way.
git-svn-id: svn://localhost/tmp/repo1/trunk@52 09e1528e-719b-4268-a937-68dd4ef2d4b4

+++ Shared/misc/cli/full-analysis.sh
@@ -6,29 +6,42 @@
 . $BASH_LIB/common.lib
 checkenv
 
+# Uncomment for debug mode.
+# DEBUG=true
+
+# Controls whether temp folder will be clened or not.
+if [ ${DEBUG} ]
+then
+    err " WARNING: Debug mode is ON."
+else
+    CLEANUP=true
+fi
+
 # ==============================
 # Local functions.
 # ==============================
 function print_usage() {
-    err "Syntax: $0 [experiment root] [social graph (stem)] [protocol name] [analysis-output-folder]"
+    err "Syntax: $0 [experiment root] [social graph (stem)] [protocol_file_stem:descriptive_name] [analysis-output-folder]"
 }
 
 function summary() {
-    VNAME="${1^^?}_SUMMARY"
+    VNAME="${1^^?}"
 
     # TODO Unify name generation with the Python part
     # (i.e. replace this part with a call to analyzer).
-    if [ ! ${2} ]
+    export $VNAME"_STEM"=$1
+
+    if [ ${2} ]
     then
-	RESULT="$1"
-    elif [ ! ${3} ]
+	export $VNAME"_SUMMARY"="${1}-summary.${2}"
+    fi	
+
+    if [ ${3} ]
     then 
-	RESULT="${1}-summary.${2}"
-    else
-	RESULT="${1}-${3}.${2}"
+	export $VNAME"_EXPERIMENT"="${1}-${3}.${2}"
     fi
+
     # ------------------------------------------------
-    export $VNAME="$RESULT"
 }
 
 function plot() {
@@ -37,13 +50,14 @@ function plot() {
     std_stats.R\
  --input ${4}/${1}\
  --metric $2\
- --algorithm $3\
- --output ${4}/${2}-${3}-linlin.eps > ${4}/${2}-stats-${3}.text
+ --algorithm "$5"\
+ --output ${4}/${2}-${3}-linlin.eps |\
+ analyzer --psyco --verbose -t python misc.cli.NumbersOnly > ${4}/${2}-stats-${3}.text
     
     std_stats.R\
  --input ${4}/${1}\
  --metric $2\
- --algorithm $3\
+ --algorithm "$5"\
  --output ${4}/${2}-${3}-loglin.eps\
  --logplot "y" > /dev/null
 } 
@@ -60,15 +74,25 @@ fi
 
 ROOT_FOLDER=$1
 SOCIAL_GRAPH=$2
-ALGORITHM=$3
+ALGORITHM=`echo $3 | awk 'BEGIN {FS=":"} {print $1}'`
+DESCRIPTIVE=`echo $3 | awk 'BEGIN {FS=":"} {print $2}'`
+
+err "Descriptive name is $DESCRIPTIVE, result stem is $ALGORITHM."
 
 # ==============================
 # Pre-analysis.
 # ==============================
 # Verifies that the folder structure is sane.
 analyzer -t python --psyco -V root=$ROOT_FOLDER experiment.util.ChkStructure
-#REPETITIONS=$?
-REPETITIONS=1
+
+REPETITIONS=$?
+
+if [ ${DEBUG} ]
+then
+    REPETITIONS=1
+fi
+
+err "Will analyse $REPETITIONS repetitions."
 
 if [ $REPETITIONS -le 0 ]
 then
@@ -80,7 +104,7 @@ assert_exists $GRAPHS_HOME/$SOCIAL_GRAPH".al"
 assert_exists $GRAPHS_HOME/$SOCIAL_GRAPH".bin"
 
 # Gets the size of the graph.
-GRAPH_SIZE=`analyzer -t python --psyco -V input=$GRAPHS_HOME/$SOCIAL_GRAPH.al misc.cli.GraphSize`
+GRAPH_SIZE=`analyzer -t python --psyco -V input=$GRAPHS_HOME/$SOCIAL_GRAPH.al graph.cli.GraphSize`
 
 # Makes a shadow copy of the root folder.
 NEWROOT="${ROOT_FOLDER}-${RANDOM}"
@@ -96,7 +120,7 @@ A=$4/$ALGORITHM
 
 mkdir -p $A/$ALGORITHM
 
-echo Verification passed.
+err Verification passed.
 
 # ==============================
 # Log parsing.
@@ -108,51 +132,66 @@ do
     # Parses the binary log.
     summary "log" "bin.gz" $i
     summary "messages" "text" $i
-    ucat $M/$LOG_SUMMARY | analyzer-j -i $GRAPHS_HOME/$SOCIAL_GRAPH".bin":stdin it.unitn.disi.analysis.LatencyComputer > $M/$MESSAGES_SUMMARY
+    ucat $M/$LOG_EXPERIMENT | analyzer-j -i $GRAPHS_HOME/$SOCIAL_GRAPH".bin":stdin it.unitn.disi.analysis.LatencyComputer > $M/$MESSAGES_EXPERIMENT
 
     # Extracts the standard set of metrics. These are:
     summary "load" "text" $i 
     summary "latency" "text" $i 
     summary "minload" "text" $i
     summary "dups" "text" $i
-
-    # 1 - load distribution;
-    # 2 - latency distribution;
-    # 3 - zero-load versus time;
-    # 4 - ideal Demers;
-    parselog $M/$MESSAGES_SUMMARY --psyco -V network_size=$GRAPH_SIZE -s\
- load:allpoints:$M/$LOAD_SUMMARY\
-,latency:allpoints:$M/$LATENCY_SUMMARY\
-,minload:allpoints:$M/$MINLOAD_SUMMARY
+    summary "total" "text" $i
     
-    # 5 - duplicates-per-message distribution.
-    analyzer --verbose -t python -V log=$M/$MESSAGES_SUMMARY:social_graph=$GRAPHS_HOME/$SOCIAL_GRAPH".al" experiment.logparse.DuplicatesPerMessage > $M/$DUPS_SUMMARY
+    # 1 - load (per-node);
+    # 2 - load (total);
+    # 3 - delivered messages (total);
+    # 4 - undelivered messages (total);
+    # 5 - latency distribution;
+    # 6 - zero-load versus time.
+    parselog --file $M/$MESSAGES_EXPERIMENT --psyco -V network_size=$GRAPH_SIZE --nolabels --statistics\
+ load:allpoints:$M/$LOAD_EXPERIMENT\
+,load:total:append:$M/$TOTAL_EXPERIMENT\
+,delivered:total:append:$M/$TOTAL_EXPERIMENT\
+,duplicates:total:append:$M/$TOTAL_EXPERIMENT\
+,undelivered:total:append:$M/$TOTAL_EXPERIMENT\
+,latency:allpoints:$M/$LATENCY_EXPERIMENT\
+,minload:allpoints:$M/$MINLOAD_EXPERIMENT
+    
+    # 8 - Duplicates-per-message (per-message).
+    analyzer --verbose -t python -V log=$M/$MESSAGES_EXPERIMENT:social_graph=$GRAPHS_HOME/$SOCIAL_GRAPH".al" experiment.logparse.DuplicatesPerMessage > $M/$DUPS_EXPERIMENT
 done
 
-# Averages load, latency, duplicates-per-message, and zero-load versus time.
-summary "load"
-summary "load_avg" "text"
-summary "latency"
-summary "latency_avg" "text"
+# Averages everything.
+summary "load" "text"    
+summary "latency" "text"
+summary "total" "text"
 
-extract_average $REPETITIONS $M/$LOAD_SUMMARY "text" > $A/$LOAD_AVG_SUMMARY
-extract_average $REPETITIONS $M/$LATENCY_SUMMARY "text" > $A/$LATENCY_AVG_SUMMARY
+extract_average $REPETITIONS $M/$LOAD_STEM "text" > $A/$LOAD_SUMMARY        # Per-node load.
+extract_average $REPETITIONS $M/$LATENCY_STEM "text" > $A/$LATENCY_SUMMARY  # Per-node latencies.
+extract_average $REPETITIONS $M/$MINLOAD_STEM "text" > $A/$MINLOAD_SUMMARY  # Zero-load versus time.
+extract_average $REPETITIONS $M/$TOTAL_STEM "text" > $A/$TOTAL_SUMMARY      # Total load, delivered, undelivered.
 
 # Now drift, and convergence figures.
 summary "drift" "text"
 summary "convergence" "text"
 summary "output"
 
-extract_average -t 'drift' $REPETITIONS "$O/$OUTPUT_SUMMARY" "text.gz" > $A/$DRIFT_SUMMARY
-extract_average -t 'STABLE' -p ':' $REPETITIONS "$O/$OUTPUT_SUMMARY" "text.gz" > $A/$CONVERGENCE_SUMMARY
+extract_average -t 'drift' $REPETITIONS "$O/$OUTPUT_STEM" "text.gz" > $A/$DRIFT_SUMMARY
+extract_average -t 'STABLE' -p ':' $REPETITIONS "$O/$OUTPUT_STEM" "text.gz" > $A/$CONVERGENCE_SUMMARY
 
 # ==============================
 # Plots graphs and statistics.
 # ==============================
 
 # Plots the graphs and prints the statistics.
+plot $LOAD_SUMMARY "load" $ALGORITHM $A "$DESCRIPTIVE"
+plot $LATENCY_SUMMARY "latency" $ALGORITHM $A "$DESCRIPTIVE"
 
-plot $LOAD_AVG_SUMMARY "load" $ALGORITHM $A
-plot $LATENCY_AVG_SUMMARY "latency" $ALGORITHM $A
+if [ ${CLEANUP} ]
+then
+    rm -rf $ROOT_FOLDER
+fi
 
-rm -rf $ROOT_FOLDER
+cd $A
+minloadplot.g
+mv minload.eps $A
+
